refactored rml mapping classes
