+++ server/server.default.properties
@@ -16,7 +16,7 @@ ontologyFile=../ontology.xml
 mappingsDir=../mappings
 
 # List of languages, e.g. 'en,de,fr' or '@mappings'
-languages=@mappings
+languages=en
 
 # We follow the same structure with the dump extraction config
